+++ dev/main-react16/src/global.jsx
@@ -52,7 +52,7 @@ const prefetchConfig = [
 
 // microApp.preFetch(prefetchConfig)
 
-window['scopeKeySpe'] = 'value from base app'
+window['escapeKey3'] = 'escapeKey3 from base app'
 window.Vue = { tip: 'Vue from base' }
 
 microApp.start({
@@ -107,7 +107,7 @@ microApp.start({
   plugins: {
     global: [
       {
-        scopeProperties: ['scopeKey1', 'scopeKey2', 'scopeKeySpe'],
+        scopeProperties: ['scopeKeyPure1', 'scopeKey1', 'scopeKey2'],
         escapeProperties: ['escapeKey1', 'escapeKey2'],
         options: {a: 1,},
         loader(code, url, options) {
@@ -118,7 +118,7 @@ microApp.start({
     ],
     modules: {
       react16: [{
-        scopeProperties: ['scopeKey3', 'scopeKey4'],
+        scopeProperties: ['scopeKeyPure2', 'scopeKey3', 'scopeKey4'],
         escapeProperties: ['escapeKey3', 'escapeKey4'],
         // loader(code, url) {
         //   if (process.env.NODE_ENV === 'development' && code.indexOf('sockjs-node') > -1) {